Fix build errors and warnings
- Fix C4099 warnings by changing struct to class in all traits forward declarations to match dynamic_graph.hpp
- Fix static assertion failures for unordered container iterators to work with both MSVC (bidirectional) and GCC (forward-only) implementations
- Fix CMake clang-debug configuration to use clang-cl instead of clang++ on Windows
- Fix narrowing conversion errors in test files for clang compiler by adding explicit static_cast

All tests now build successfully on MSVC and Clang on Windows.

+++ tests/test_dynamic_graph_cpo_dos.cpp
@@ -1192,7 +1192,7 @@ TEST_CASE("dos CPO integration", "[dynamic_graph][dos][cpo][integration]") {
         visited[start] = true;
         order.push_back(start);
         
-        std::vector<uint32_t> queue = {start};
+        std::vector<uint32_t> queue = {static_cast<uint32_t>(start)};
         while (!queue.empty()) {
             auto uid = queue.front();
             queue.erase(queue.begin());
