+++ tests/test_dynamic_graph_cpo_dous.cpp
@@ -1198,7 +1198,7 @@ TEST_CASE("vos CPO integration", "[dynamic_graph][dous][cpo][integration]") {
         visited[start] = true;
         order.push_back(start);
         
-        std::deque<uint32_t> queue = {start};
+        std::deque<uint32_t> queue = {static_cast<uint32_t>(start)};
         while (!queue.empty()) {
             auto uid = queue.front();
             queue.erase(queue.begin());
